@@ -4,10 +4,10 @@
   "private": true,
   "dependencies": {
     "bootstrap": "^5.3.0",
-    "react": "16.11.0",
-    "react-dom": "16.11.0",
-    "react-scripts": "3.2.0",
-    "web3": "1.2.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1",
+    "web3": "^1.2.2"
   },
   "scripts": {
     "start": "react-scripts --openssl-legacy-provider start",
